daemon/network: Deprecate dated fields from `Settings`
This struct holds the overall networking state of a container, and still
have a bunch of relic of the past (ie. < Docker v1.09).

All the following fields are never written and are only used to create
the equivalent fields in the api `NetworkSettingsBase` (which was
properly deprecated in ce574949f):

- `HairpinMode`
- `LinkLocalIPv6Address`
- `LinkLocalIPv6PrefixLen`
- `SecondaryIPAddresses`
- `SecondaryIPv6Addresses`

As such, all those fields are now marked as deprecated and their values
aren't copied into `NetworkSettingsBase` anymore.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>