libnetwork/networkdb: unify Event-types: export Event and add Type field
All events generated share the same underlying (event) type, but had to
be type-matched to detect the kind of event (create, update, delete).

Instead of defining different types for each event, we can share the
same type, and instead add a Type field, which holds the kind of event.

This patch:

- Adds a Type field to the event-type and exports it. The Type field
  uses the driverapi.EventType.
- Replaces uses of CreateEvent, UpdateEvent and DeleteEvent to use the
  Event type instead, and setting the Event.Type field accordingly.

Type-matching is still needed, due to the "go-events" package defining
the `Event` type as a generic `interface{}`, but logging was added for
cases where we get an unexpected type.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>