daemon/info: add BuildKit version
This is a very simple first implementation; it relies on the way that
`bkversion.Version` is overwritten in an initializer, and it doesn't
expose the built-in Dockerfile frontend version/revision as that
information is not currently available to BuildKit.

That being said, this is still a useful debug tool as-is given the
increasing importance of the integrated BuildKit version for many users.

Signed-off-by: Bjorn Neergaard <bjorn.neergaard@docker.com>