libnetwork: fan out firewall reload top-down
The pattern of having low-level code that sets up iptables rules
register reload callbacks for itself is not ideal. Callbacks can not be
unregistered, so any values captured by the closures can never be GC'ed.
And the callback registry is a singleton. While these issues could be
overcome by iterating on the API, there are some other insurmountable
issues which cannot be overcome without a complete overhaul of how the
firewall reload logic is plumbed through. Imperative callback
registration makes the control flow hard to reason about and the order
of callbacks is implied by the order they were registered.

Make the libnetwork Controller ultimately responsible for replaying
firewall configuration when the firewall rulesets have been flushed
externally. It fans out the reload signal to the drivers, which in turn
fan out to the individual networks.

Signed-off-by: Cory Snider <csnider@mirantis.com>