DNM: PoC Always set a userns
This makes it so all containers, unless specically requesting the host
namespace with `--privileged` or `--userns=host`, are placed into a new
user namespace. UIDs and GIDs are mapped 1:1, so uid 0 in the userns is
uid 0 on the host.

The main thing this buys us is the containerized process no longer has
privileges in the root userns which simultanously makes it lose any root
namespace privileges and also allows us to safely grant (barring bugs
and oversights, of course) extra privileges to the container.

Specifically what's nice about this is now a containerized process can
do bind and fuse mounts, and possibly overlay depending on which kernel
you are running (ubuntu kernel allows unprivileged overlay mounts).

This *should* actually be more secure than allowing the container to run
in the host namespace, and even if we don't grant CAP_SYS_ADMIN in the
usernamespace by default I think this would be a good change to make
because the container process should no longer have any capabilities in
the root namespace.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>