container: Remove RWLayer
Don't store the RWLayer in the Container object - it's an implementation
detail of the graphdriver ImageService and it shouldn't be exposed.
The ID of the layer is always the same as the container ID so it can be
used to move the direct RWLayer accesses to the graphdriver's
ImageService implementation.

Signed-off-by: Paweł Gronowski <pawel.gronowski@docker.com>